Move reaction test to its own file
