+++ testing/matrix-sdk-integration-testing/src/tests.rs
@@ -1,3 +1,4 @@
 mod invitations;
+mod reactions;
 mod redaction;
 mod repeated_join;
